requirements.txt
python-dotenv==1.1.0
PyYAML==6.0.2

langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.18
langchain-groq==0.3.2        
langchain-huggingface==0.2.0 
langchain-community==0.3.24
langchain-text-splitters==0.3.8

chromadb==1.0.13
chroma-hnswlib==0.7.6

tiktoken==0.9.0

sentence-transformers==2.7.0
torch==2.3.1
numpy==1.26.4